Initial Commit - Basic Prototype
Application has basic Client and a Server, that communicate. Server
sends JPanel to the the Client and client repaints the screen. This
basic functionality will enable us to create different sorts of panels
and pass it to the client for guessing. The lower panel (the actual
challege) will be passed in the same way with a panel.